test(ui-rewrite): consolidate duplicate test files and fix tsc build errors
